copy.sh will now check if git files already exist

+++ copy.sh
@@ -1,8 +1,8 @@
 #!/bin/sh
-cp git-completion.bash //Library/Developer/CommandLineTools/usr/share/git-core/git-completion.bash
-cp git-prompt.sh /Library/Developer/CommandLineTools/usr/share/git-core/git-prompt.sh
-cp .bash_profile ~/.bash_profile
+[[ -e /Library/Developer/CommandLineTools/usr/share/git-core/git-completion.bash ]] || cp git-completion.bash /Library/Developer/CommandLineTools/usr/share/git-core/git-completion.bash
+[[ -e /Library/Developer/CommandLineTools/usr/share/git-core/git-prompt.sh ]] || cp git-prompt.sh /Library/Developer/CommandLineTools/usr/share/git-core/git-prompt.sh
+cp -f .bash_profile ~/.bash_profile
 source ~/.bash_profile
 
 
-# Comment to line 2 and 3 : You need to edit that script to check for existence of this files and if they don't exist then copy them to this directories!!
+# Comment to line 2 and 3: Part before || checks if this files already exists. If they do the second part won't be executed. If they don't the second part WILL be executed.
